requirements.txt
docling==2.62.0
docling-core==2.51.1
docling-ibm-models==3.10.2
docling-parse==4.7.1
Pillow==11.3.0
torch==2.9.1
torchvision==0.24.1
transformers==4.57.1
pydantic==2.12.4
numpy==2.2.6
huggingface-hub==0.36.0